requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
asyncpg==0.30.0
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
aiosmtplib==4.0.2
bcrypt==4.3.0
python-jose==3.5.0
pydantic==2.11.7



